Refactor temperature monitors
Make a get_temperatures() method be the abstract method that child classes override rather than send_temperatures(), whose implementation can live in the parent class.
