+++ finesse/hardware/temperature/dummy_temperature_monitor.py
@@ -22,15 +22,8 @@ class DummyTemperatureMonitor(TemperatureMonitorBase):
         pub.sendMessage("temperature_monitor.close")
         logging.info("Closed connection to dummy temperature monitor")
 
-    def send_temperatures(self) -> None:
-        """Publish fluctuating temperatures."""
+    def get_temperatures(self) -> list[Decimal]:
+        """Get current temperatures."""
         _BASE_TEMPS = (19, 17, 26, 22, 24, 68, 69, 24)
         noise = self._noise_producer()
-        temperatures = [Decimal(noise + temp) for temp in _BASE_TEMPS]
-        time_now = datetime.now().timestamp()
-
-        pub.sendMessage(
-            "temperature_monitor.data.response",
-            temperatures=temperatures,
-            time=time_now,
-        )
+        return [Decimal(noise + temp) for temp in _BASE_TEMPS]
